Tested with Dexie.js
Successfully runs with all unit tests for Dexie on IE11.

This is my initial commit of this polyfill. It still lacks checks
whether it should be included or not (check if browser is IE10/IE11). It
also could benefit from some comment removings and a set of unit tests
of its own.