Support for compound primary key + bugfixes
Added support for not only compound indexes but also compound primary
keys, so that the W3C web-platform-test "idbcursor_iterating_index.htm"
is passed.

Also fixes the missing property IDBObjectStore.autoIncrement (not
visible in IE. But now it is).

Also found some bugs in the polyfill:
* onsuccess never got called from IDBObjectStore.add() when there was a
compound or multiValue index.
* get() and count() methods didnt accept IDBKeyRange. Only direct keys.
* Missing meta-data cleanup in deleteObjectStore() and deleteIndex().

The W3C web-platform-tests that applies to all the features of this
polyfill now works. These are:
* idbindex-multientry-arraykeypath.htm
* idbindex-multientry-big.htm
* idbindex-multientry.htm
* idbcursor_iterating.htm
* idbcursor_iterating_index.htm
* idbcursor_iterating_index2.htm