Found out that I need to implement own handling of autoIncrement in case we intercept the objectStore.
Implented that and swapped the order of performMainOperations() with performMetaOperations() so that metaOperations is always executed first. THis is important because if any metaOperation fails, we will need to make the main operation fail. For example if key is unique, a violation of that will only fail in the metaOperation. Also if out of storage, the same thing applies.

Now almost everything is finished. Just to complete the add/put/delete/clear methods is left.